backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
google-generativeai
chromadb
beautifulsoup4
playwright
requests
python-dotenv
pydantic
aiohttp
markdown
plotly
pandas
numpy
scikit-learn
scipy
PyPDF2
pytesseract
Pillow
python-multipart
sse-starlette
lxml